# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Learning System API",
    "description": "Backend API for Online Learning Platform - Challenges 1-3\n\n## Challenges Completed:\n- **Challenge 1**: Database Design (PostgreSQL, 7 tables, 3NF normalized)\n- **Challenge 2**: CRUD Operations (Courses & Classes)\n- **Challenge 3**: Authentication (Register, Login, JWT, Profile)\n\n## Features:\n- ✅ User Registration & Login with JWT\n- ✅ Role-based Authorization (admin, instructor, student)\n- ✅ Course Management (CRUD with instructor/admin)\n- ✅ Class Management (CRUD with pagination)\n- ✅ Pagination & Filtering\n- ✅ Error Handling & Validation\n- ✅ Logging & Monitoring\n\n## Test Credentials:\n- **Email**: student1@example.com\n- **Password**: password123\n- **Role**: student",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/courses": {
      "get": {
        "tags": ["Courses"],
        "summary": "List Courses",
        "description": "Get list of all courses with pagination and filtering",
        "operationId": "listCourses",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["draft", "published"]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "courses": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "title": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              },
                              "instructor": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "integer"
                                  },
                                  "name": {
                                    "type": "string"
                                  },
                                  "email": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Courses"],
        "summary": "Create Course",
        "description": "Create a new course (instructor/admin only)",
        "operationId": "createCourse",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Advanced Python Programming"
                  },
                  "description": {
                    "type": "string",
                    "example": "Master advanced Python concepts"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published"],
                    "default": "draft"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "tags": ["Courses"],
        "summary": "Get Course by ID",
        "operationId": "getCourseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course retrieved"
          },
          "404": {
            "description": "Course not found"
          }
        }
      },
      "put": {
        "tags": ["Courses"],
        "summary": "Update Course",
        "operationId": "updateCourse",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "published"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "delete": {
        "tags": ["Courses"],
        "summary": "Delete Course",
        "operationId": "deleteCourse",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token received from /api/auth/login endpoint"
      }
    }
  }
}